requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.4.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data processing and graph analysis
networkx>=3.1
imbalanced-learn>=0.11.0

# Visualization and dashboard
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0

# Model interpretability
captum>=0.6.0

# Utilities and configuration
tqdm>=4.65.0
pyyaml>=6.0
loguru>=0.7.0
hydra-core>=1.3.0

# Experiment tracking
wandb>=0.15.0

# Development and testing
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
